admin: extract license utils
Extract it to a utils file to allow sharing it with other admin
endpoints as well later.
