+++ src/v/redpanda/admin/services/utils.h
@@ -1,0 +1,20 @@
+/*
+ * Copyright 2025 Redpanda Data, Inc.
+ *
+ * Use of this software is governed by the Business Source License
+ * included in the file licenses/BSL.md
+ *
+ * As of the Change Date specified in that file, in accordance with
+ * the Business Source License, use of this software will be governed
+ * by the Apache License, Version 2.0
+ */
+
+#pragma once
+
+#include "features/feature_table.h"
+
+namespace admin::utils {
+
+void check_license(const features::feature_table& ft);
+
+} // namespace admin::utils
